refactor: create room 분리
