+++ apps/media/src/mediasoup/mediasoup.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 
 import { MediasoupService } from './mediasoup.service';
+import { RoomService } from 'src/room/room.service';
+import { RoomModule } from 'src/room/room.module';
 
 @Module({
-  imports: [],
+  imports: [RoomModule],
   providers: [MediasoupService],
   exports: [MediasoupService],
 })
