+++ apps/media/src/room/room.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { RoomService } from './room.service';
+
+@Module({
+  providers: [RoomService],
+  exports: [RoomService],
+})
+export class RoomModule {}
